pyproject.toml
[project]
name = "larper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "textual>=8.2.3",
    "watchdog>=6.0.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=2.7.0",
    "pytest-asyncio>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
